all uint64s changed to double

+++ pickANumberExercise/pickANumberExercise/Program.cs
@@ -12,17 +12,17 @@ namespace pickANumberExercise
         {
             Console.WriteLine("Write any number: ");
             string numResponse = Console.ReadLine();
-            UInt64 inputNum = Convert.ToUInt64(numResponse);
-            UInt64 multiply = inputNum * 50;
+            double inputNum = Convert.ToDouble(numResponse);
+            double multiply = inputNum * 50.0;
             Console.WriteLine(inputNum + " multiplied by 50 is " + multiply);
-            UInt64 add = inputNum + 25;
+            double add = inputNum + 25.0;
             Console.WriteLine(inputNum + " added to 25 is " + add);
             double divide = inputNum / 12.5;
             Console.WriteLine(inputNum + "divided by 12.5 is " + divide);
             bool compare = inputNum > 50;
             Console.WriteLine(inputNum + " is greater than 50?");
             Console.WriteLine(compare);
-            UInt64 remainder = inputNum % 7;
+            double remainder = inputNum % 7.0;
             Console.WriteLine(inputNum + " divided by 7 has a remainder of " + remainder);
             Console.Read();
         }
